Add a few fixes (one is prefix sbnana/ in a few includes in NueCuts.h, seems there are more in the repo, and another changing the type int to double for best plane dEdx variable in NueVars)

+++ sbnana/SBNAna/Vars/NueVars.cxx
@@ -45,7 +45,7 @@ const Var kRecoShower_TruePdg(
 
 // Currently assumes shw 0 is the primary
 const Var kRecoShower_BestdEdx(
-    [](const caf::SRSliceProxy* slc) -> int {
+    [](const caf::SRSliceProxy* slc) -> double {
       double dedx = -5.0;
       const int largestShwIdx(kLargestRecoShowerIdx(slc));
       if (largestShwIdx != -1 && slc->reco.shw[largestShwIdx].bestplane_dEdx > 0) {
